extract test functions to htmlConstants
